vertex-deploy/training/requirements.txt
numpy
scipy
scikit-learn
joblib
google-cloud-storage